Add setting for deleting etherpads on the server on Etherpad obj delete

+++ cosinnus_etherpad/models.py
@@ -170,10 +170,12 @@ def delete_etherpad_group(sender, instance, **kwargs):
     """
     Receiver to delete a group on etherpad server
     """
-    client = _init_client()
-    group_id = client.createGroupIfNotExistsFor(
-        groupMapper=_get_group_mapping(instance))
-    client.deleteGroup(groupID=group_id['groupID'])
+    from cosinnus.conf import settings as cosinnus_settings
+    if getattr(cosinnus_settings, 'COSINNUS_DELETE_ETHERPADS_ON_SERVER_ON_DELETE', False):
+        client = _init_client()
+        group_id = client.createGroupIfNotExistsFor(
+            groupMapper=_get_group_mapping(instance))
+        client.deleteGroup(groupID=group_id['groupID'])
 
 
 @receiver(pre_save, sender=Etherpad)
@@ -199,13 +201,15 @@ def delete_etherpad(sender, instance, **kwargs):
     """
     Receiver to delete a pad on etherpad server
     """
-    if not instance.is_container:
-        try:
-            instance.client.deletePad(padID=instance.pad_id)
-        except EtherpadException as exc:
-            # failed deletion of missing padIDs is ok
-            if 'padID does not exist' not in str(exc):
-                raise
+    from cosinnus.conf import settings as cosinnus_settings
+    if getattr(cosinnus_settings, 'COSINNUS_DELETE_ETHERPADS_ON_SERVER_ON_DELETE', False):
+        if not instance.is_container:
+            try:
+                instance.client.deletePad(padID=instance.pad_id)
+            except EtherpadException as exc:
+                # failed deletion of missing padIDs is ok
+                if 'padID does not exist' not in str(exc):
+                    raise
 
 def _get_group_mapping(group):
     return smart_text('p_%d_g_%s' % (CosinnusPortal.get_current().id, group.slug)).encode('utf-8')
